Split brush source enclosure diagnostics
